third_party/blink/web_tests/platform/generic/fast/forms/color/input-color-onchange-event-expected.txt

Test if change event fires properly when color chooser changes. Bug 66848
To manually test this, click on the input color element in the top left corner and change the value from the color chooser. See if the number of "value changed" messages matches the number of times you changed the color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.selectColorInColorChooser({}, '#ff0000'); threw exception TypeError: Failed to execute 'selectColorInColorChooser' on 'Internals': parameter 1 is not of type 'Element'..
PASS internals.selectColorInColorChooser(document, '#ff0000'); threw exception TypeError: Failed to execute 'selectColorInColorChooser' on 'Internals': parameter 1 is not of type 'Element'..
input event dispatched - value is:  #ff0000
PASS input.value is "#ff0000"
change event dispatched - value changed to #ff0000
PASS input.value is "#ff0000"
Change event is only dispatched, when color chooser is closed
input event dispatched - value is:  #ff0002
PASS Change event was dispatched.
PASS successfullyParsed is true

TEST COMPLETE
